requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
pandas
openpyxl
pdf2image
Pillow
sentence_transformers
faiss-cpu
torch 
torchvision 
torchaudio
numpy
pdfminer.six
PyMuPDF
pytesseract
pdfplumber
